+++ CHANGELOG.md
@@ -38,6 +38,7 @@ _Note: Gaps between patch versions are faulty, broken or test releases._
 * Improved `Object.fromArray` API
 * Added new parameters for `core/request/interface/RequestAPI`
 * Improved date/number formatting API
+* Improved API of `prelude/string`
 
 #### :house: Internal